refactor: simplify entry points and remove unused loader file in repack-plugin
