+++ package.json
@@ -52,10 +52,6 @@
       "types": "./repack-plugin/index.d.ts",
       "default": "./repack-plugin/index.js"
     },
-    "./repack-plugin/loader.js": {
-      "import": "./repack-plugin/loader.js",
-      "default": "./repack-plugin/loader.js"
-    },
     "./package.json": "./package.json",
     "./server": {
       "types": "./lib/typescript/src/server/index.d.ts",
