+++ repack-plugin/src/loader.ts
@@ -1,36 +1,36 @@
-import { transform } from '@babel/core';
-import type { LoaderContext } from '@rspack/core';
-import type { UnistylesPluginOptions } from '../../plugin';
-import { REACT_NATIVE_COMPONENT_NAMES, REPLACE_WITH_UNISTYLES_PATHS } from '../../plugin/src/consts';
+import { transform } from "@babel/core"
+import type { LoaderContext } from "@rspack/core"
+import type { UnistylesPluginOptions } from "../../plugin"
+import { REACT_NATIVE_COMPONENT_NAMES, REPLACE_WITH_UNISTYLES_PATHS } from "../../plugin/src/consts"
 
-const importName = 'react-native-unistyles';
+const importName = "react-native-unistyles"
 
 interface UnistylesLoaderOptions {
-    babelPlugins?: string[];
-    unistylesPluginOptions?: UnistylesPluginOptions;
+    babelPlugins?: string[]
+    unistylesPluginOptions?: UnistylesPluginOptions
 }
 
 const UNISTYLES_REGEX = new RegExp(
-    [...REACT_NATIVE_COMPONENT_NAMES, ...REPLACE_WITH_UNISTYLES_PATHS, importName].join('|')
-);
+    [...REACT_NATIVE_COMPONENT_NAMES, ...REPLACE_WITH_UNISTYLES_PATHS, importName].join("|")
+)
 
-export default function unistylesLoader(this: LoaderContext<UnistylesLoaderOptions>, source: string) {
-    this.cacheable();
-    const callback = this.async();
-    const options = this.getOptions();
+export function unistylesLoader(this: LoaderContext<UnistylesLoaderOptions>, source: string) {
+    this.cacheable()
+    const callback = this.async()
+    const options = this.getOptions()
 
     if (!UNISTYLES_REGEX.test(source)) {
-        callback(null, source);
-        return;
+        callback(null, source)
+        return
     }
 
-    const unistylesOptions = options.unistylesPluginOptions;
+    const unistylesOptions = options.unistylesPluginOptions
 
     const unistylesPlugin = unistylesOptions
-        ? ['react-native-unistyles/plugin', unistylesOptions]
-        : 'react-native-unistyles/plugin';
+        ? ["react-native-unistyles/plugin", unistylesOptions]
+        : "react-native-unistyles/plugin"
 
-    const babelPlugins = options.babelPlugins ?? [];
+    const babelPlugins = options.babelPlugins ?? []
 
     transform(
         source,
@@ -44,14 +44,13 @@ export default function unistylesLoader(this: LoaderContext<UnistylesLoaderOptio
         },
         (err, result) => {
             if (err) {
-                callback(err);
-                return;
+                callback(err)
+                return
             }
 
             //@ts-ignore
-            callback(null, result.code, result.map);
-
-            return;
+            callback(null, result.code, result.map)
+            return
         }
-    );
+    )
 }
